Moved implementation to header for BOB
BoBshield implementation moved purely to a header file. This is to prevent collision with the ISR for the Servo and for the non-servo examples.
