+++ examples/BOBShield/BOBShield_SelfTest/BOBShield_SelfTest.ino
@@ -1,5 +1,6 @@
-#include "BOBShield.h"
- 
+#include <BOBShield.h>
+#include <SamplingServo.h>
+
   void wait() {
   delay(5000);
 }
